Consolidated ezobjects-mysql with ezobjects, identifying based on tableName
